Use HttParty as transport layer
Seems like tfl is now returning a lower case content type header,
however curb is hardcoding "Content-Type" when trying to parse it (which
in turn makes the whole thing die). Switching to HttParty seem to fix it
as it will not make any assumptions on the underlying combination of
upper and lower case letters.