Execute jobs sequentially
CircleCI seems to have added a limit to how many concurrent jobs we can
have, without however having an easy way to control that concurrency.
This adds a dependency on the code_quality step so that the workflow is
in fact executed sequentially. Longer term, it feels like we should move
away from CircleCI - we can live with the sequentiality as this is a low
traffic project, but it should be easy to control that (or should be
controlled by Circle itself as they know what plan we're on).